Fix inode->xcache leak. Add full leak check and fix xattr test in Makefile. Fix various compiler warnings.

+++ user/test/tux3.c
@@ -193,7 +193,7 @@ int main(int argc, const char *argv[])
 		if (!name)
 			goto usage;
 		if (!strcmp(command, "get")) {
-			printf("read xattr %.*s\n", strlen(name), name);
+			printf("read xattr %.*s\n", (int)strlen(name), name);
 			struct xattr *xattr = get_xattr(inode, name, strlen(name));
 			if (!xattr) {
 				errno = EINVAL;
